hikikomori-bot/requirements.txt
tweepy==4.14.0
openai==1.14.0
python-dotenv==1.0.1
schedule==1.2.1
